requirements.txt
# Core web UI
streamlit>=1.25.0

# Data processing
pandas>=2.1.0
numpy>=1.26.0

# Visualization
plotly>=5.15.0
altair>=5.6.0

# HTTP / APIs
requests>=2.31.0

# Google Gemini / Generative AI client
google-generativeai>=0.3.0
google-api-core>=2.11.0

# Helpful utilities
python-dotenv>=1.1.0

# Geospatial libraries
rasterio>=1.3.6
geopandas>=0.14.0
fiona>=1.9.4
pyproj>=3.6.0
gdal>=3.7.4
shapely>=2.1.0
geographiclib>=2.1 
geopy>=2.4.1